Moved coined transitions into their own subproject. fixes TweetWallFX/TweetwallFX#505
