+++ controls/build.gradle
@@ -25,4 +25,5 @@
 dependencies {
     compile project(':TweetWallFX-Core')
     compile project(':TweetWallFX-StepEngine-DataProviders')
+    compile project(':TweetWallFX-Transitions')
 }
