+++ controls/src/main/java/org/tweetwallfx/controls/steps/ImageMosaicStep.java
@@ -44,14 +44,14 @@ import javafx.scene.image.ImageView;
 import javafx.scene.layout.Pane;
 import javafx.util.Duration;
 import org.tweetwallfx.controls.WordleSkin;
-import org.tweetwallfx.controls.transition.LocationTransition;
-import org.tweetwallfx.controls.transition.SizeTransition;
 import org.tweetwallfx.stepengine.api.DataProvider;
 import org.tweetwallfx.stepengine.api.Step;
 import org.tweetwallfx.stepengine.api.StepEngine.MachineContext;
 import org.tweetwallfx.stepengine.api.config.StepEngineSettings;
 import org.tweetwallfx.stepengine.dataproviders.ImageMosaicDataProvider;
 import org.tweetwallfx.stepengine.dataproviders.ImageMosaicDataProvider.ImageStore;
+import org.tweetwallfx.transitions.LocationTransition;
+import org.tweetwallfx.transitions.SizeTransition;
 
 public class ImageMosaicStep implements Step {
