+++ controls/src/main/java/org/tweetwallfx/controls/steps/NodeFadeOutStep.java
@@ -27,10 +27,10 @@ import java.util.Objects;
 import javafx.scene.Node;
 import static org.tweetwall.util.ToString.*;
 import org.tweetwallfx.controls.WordleSkin;
-import org.tweetwallfx.controls.transition.FlipOutXTransition;
 import org.tweetwallfx.stepengine.api.Step;
 import org.tweetwallfx.stepengine.api.StepEngine;
 import org.tweetwallfx.stepengine.api.config.StepEngineSettings;
+import org.tweetwallfx.transitions.FlipOutXTransition;
 
 /**
  * Configrable step that fades out a {@link Node} lookup up via its
