+++ devoxx-2017-be/cinema/src/main/java/org/tweetwallfx/devoxx17be/steps/Devoxx17FlipOutTweets.java
@@ -32,10 +32,10 @@ import javafx.scene.Node;
 import javafx.scene.layout.VBox;
 import javafx.util.Duration;
 import org.tweetwallfx.controls.WordleSkin;
-import org.tweetwallfx.controls.transition.FlipOutXTransition;
 import org.tweetwallfx.stepengine.api.Step;
 import org.tweetwallfx.stepengine.api.StepEngine.MachineContext;
 import org.tweetwallfx.stepengine.api.config.StepEngineSettings;
+import org.tweetwallfx.transitions.FlipOutXTransition;
 
 /**
  * Devox 2017 TweetStream Flip Out Animation Step
