+++ devoxx-2017-be/exhibition/src/main/java/org/tweetwallfx/devoxx17be/exhibition/Devoxx17ShowSchedule.java
@@ -42,13 +42,13 @@ import javafx.scene.text.Text;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.tweetwallfx.controls.WordleSkin;
-import org.tweetwallfx.controls.transition.FlipInXTransition;
 import org.tweetwallfx.devoxx.cfp.stepengine.dataprovider.ScheduleDataProvider;
 import org.tweetwallfx.devoxx.cfp.stepengine.dataprovider.SessionData;
 import org.tweetwallfx.stepengine.api.DataProvider;
 import org.tweetwallfx.stepengine.api.Step;
 import org.tweetwallfx.stepengine.api.StepEngine.MachineContext;
 import org.tweetwallfx.stepengine.api.config.StepEngineSettings;
+import org.tweetwallfx.transitions.FlipInXTransition;
 
 /**
  * Devox 2017 Show Schedule (Flip In) Animation Step
