+++ voxxed-days-zurich/src/main/java/org/tweetwallfx/vdz/steps/ShowSchedule.java
@@ -41,12 +41,12 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.tweetwallfx.controls.WordleSkin;
 import org.tweetwallfx.stepengine.api.DataProvider;
-import org.tweetwallfx.controls.transition.FlipInXTransition;
 import org.tweetwallfx.devoxx.cfp.stepengine.dataprovider.ScheduleDataProvider;
 import org.tweetwallfx.devoxx.cfp.stepengine.dataprovider.SessionData;
 import org.tweetwallfx.stepengine.api.Step;
 import org.tweetwallfx.stepengine.api.StepEngine.MachineContext;
 import org.tweetwallfx.stepengine.api.config.StepEngineSettings;
+import org.tweetwallfx.transitions.FlipInXTransition;
 
 /**
  * Devox 2017 Show Schedule (Flip In) Animation Step
